+++ src/prj_profile.py
@@ -8,11 +8,13 @@ class Profile:
     CONFIG_FILE_NAME: str = "boomer.ini"
 
     name: str
+    bom_has_column_headers: bool
     bom_first_row: int # 0-based
     bom_last_row: int # 0-based
     bom_separator: str
     bom_designator_col: str
     bom_comment_col: str
+    pnp_has_column_headers: bool
     pnp_first_row: int # 0-based
     pnp_last_row: int # 0-based
     pnp_separator: str
@@ -22,16 +24,21 @@ class Profile:
 
     def __init__(self, cfgparser: configparser.ConfigParser):
         self.name = "initial-profile"
+
+        self.bom_has_column_headers = True
         self.bom_first_row = 0
         self.bom_last_row = -1 # not saved in config file
         self.bom_separator = "COMMA"
         self.bom_designator_col = "?"
         self.bom_comment_col = "?"
+
+        self.pnp_has_column_headers = True
         self.pnp_first_row = 0
         self.pnp_last_row = -1 # not saved in config file
         self.pnp_separator = "COMMA"
         self.pnp_designator_col = "?"
         self.pnp_comment_col = "?"
+
         self.__config = cfgparser
 
     def load(self, name: str):
@@ -41,17 +48,26 @@ class Profile:
             if self.__config.has_section(f'profile.{self.name}'):
                 section = self.__config[f'profile.{self.name}']
 
+                self.bom_has_column_headers = section.get("bom_has_column_headers", "True") == "True"
                 self.bom_first_row = int(section.get("bom_first_row", "0"))
                 self.bom_last_row = -1
                 self.bom_separator = section.get("bom_separator", "COMMA")
                 self.bom_designator_col = section.get("bom_designator_col", "?")
                 self.bom_comment_col = section.get("bom_comment_col", "?")
+                if self.bom_has_column_headers == False:
+                    self.bom_designator_col = int(self.bom_designator_col)
+                    self.bom_comment_col = int(self.bom_comment_col)
 
+                self.pnp_has_column_headers = section.get("pnp_has_column_headers", "True") == "True"
                 self.pnp_first_row = int(section.get("pnp_first_row", "0"))
                 self.pnp_last_row = -1
                 self.pnp_separator = section.get("pnp_separator", "COMMA")
                 self.pnp_designator_col = section.get("pnp_designator_col", "?")
                 self.pnp_comment_col = section.get("pnp_comment_col", "?")
+                if self.pnp_has_column_headers == False:
+                    self.pnp_designator_col = int(self.pnp_designator_col)
+                    self.pnp_comment_col = int(self.pnp_comment_col)
+
             else:
                 logging.warning(f"No section {self.name} in config file")
         else:
@@ -60,10 +76,12 @@ class Profile:
     def save(self):
         logging.info(f"Save profile: {self.name}")
         self.__config[f"profile.{self.name}"] = {
+            "bom_has_column_headers": self.bom_has_column_headers,
             "bom_first_row": self.bom_first_row,
             "bom_separator": self.bom_separator,
             "bom_designator_col": self.bom_designator_col,
             "bom_comment_col": self.bom_comment_col,
+            "pnp_has_column_headers": self.pnp_has_column_headers,
             "pnp_first_row": self.pnp_first_row,
             "pnp_separator": self.pnp_separator,
             "pnp_designator_col": self.pnp_designator_col,